+++ crates/core/machine/src/operations/or.rs
@@ -13,7 +13,7 @@ pub struct OrOperation<T> {
 }
 
 impl<F: Field> OrOperation<F> {
-    pub fn populate(&mut self, record: &mut ExecutionRecord, x: u32, y: u32) -> u32 {
+    pub fn populate(&mut self, record: &mut ExecutionRecord, x: u64, y: u64) -> u64 {
         let expected = x | y;
         let x_bytes = x.to_le_bytes();
         let y_bytes = y.to_le_bytes();
